+++ transport/stdio/process_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package stdio
+
+import "os/exec"
+
+func configureProcessCommand(*exec.Cmd) {}